refactor: Merge ZoomableImage into PinchToZoom
